pyproject.toml
[project]
name = "eco-damage-monitor"
version = "0.1.0"
description = "Compliant public-web monitoring and analysis system for wide-area ecological damage information."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "click>=8.1.7",
  "httpx>=0.27.0",
  "lxml>=5.2.1",
  "loguru>=0.7.2",
  "pandas>=2.2.2",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1",
  "pyyaml>=6.0.1",
  "python-dateutil>=2.9.0.post0",
  "rapidfuzz>=3.9.0",
  "scikit-learn>=1.4.2",
  "simhash>=2.1.2",
  "sqlalchemy>=2.0.29",
  "streamlit>=1.34.0",
  "tenacity>=8.3.0",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dynamic = ["playwright>=1.43.0"]
ml = ["sentence-transformers>=2.7.0", "bertopic>=0.16.0"]
dev = ["pytest>=8.2.0", "pytest-cov>=5.0.0"]

[project.scripts]
eco-monitor = "eco_damage_monitor.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
